feat: Enhance sales entry form with quantity type toggle and validation
- Added support for quantity type selection (units/weight) in the sales entry form.
- Implemented dynamic validation for quantity input based on selected type.
- Updated price input field to show error messages only after user interaction.
- Improved UI consistency by applying Mix tokens for colors and styles across components.
- Refactored sales entry form layout for better user experience and accessibility.
- Introduced helper text and hints for quantity and price fields.
- Enhanced profit display box visibility based on form validity.
- Added tests for product quantity type handling and validation logic.
- Documented UI/UX research findings for daily sales interface.

+++ lib/features/sales/data/models/sale_item_model.dart
@@ -17,8 +17,8 @@ class SaleItemModel {
   /// Product barcode (snapshot at time of sale)
   String? barcode;
 
-  /// Quantity purchased in this transaction
-  late int quantity;
+  /// Quantity purchased in this transaction (int for units, double for weight)
+  late double quantity;
 
   /// Unit selling price at time of purchase
   late double unitPrice;
@@ -43,7 +43,7 @@ class SaleItemModel {
       ..productId = entity.product.id
       ..productName = entity.product.name
       ..barcode = entity.product.barcode
-      ..quantity = entity.quantity
+      ..quantity = entity.quantity.toDouble()
       ..unitPrice = entity.unitPrice
       ..costPrice = entity.product.costPrice;
     return model;
